logproto: refactor log_proto_server_validate_options() method
Earlier the root-level validate_options implementation was delegated
to LogProtoServerOptions, which is not how it was invoked (associated with
LogProtoServer, not its Options struct).

Also, since the encoding() option is now being validated at parsing time,
the core of the original validate_options implementation is gone.

So this patch moves log_proto_server_validate_options() to LogProtoServer,
with an empty initial implementation that can be invoked by
derived classes.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>