light-common: add BackgroundEventLoop
BackgroundEventLoop is singleton async event loop that is running in
a separate background thread.

Async tasks/jobs can be scheduled by calling run_async(), which returns
a Future.

wait_async_result() can be used to block the caller thread until the
async job is completed. It returns the job's result or raises a
TimeoutError.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>