light-common: add TCPServer and UnixStreamServer
SingleConnectionStreamServer is an abstract class that implements some IO
functionality on stream-based servers, using only one connection.

The following public methods are available:
- start(), stop(), close_client()
- readexactly(n), readuntil(separator), readline(), write(data)

SingleConnectionTCPServer is a basic TCP/TLS server implementation based
on SingleConnectionStreamServer.

UnixStreamServer is the same, but it operates on Unix domain stream
sockets.

Note: This is a low-level network layer, it operates on bytes objects,
not on strings or messages.

Note: Using async and asyncio was not an essential decision, blocking on
each async call has no advantage over a blocking threaded
implementation. async is used here because the streaming and thread
synchronization APIs of asyncio are elegant and easier to understand in
my opinion.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>