light-common: add UDPServer and UnixDatagramServer
Public interface:
- start(), stop()
- read_dgram(), write_dgram(data)

Note: This is a low-level network layer, it operates on bytes objects,
not on strings or messages.

Note: Using async and asyncio was not an essential decision, blocking on
each async call has no advantage over a blocking threaded
implementation. async is used here because the streaming and thread
synchronization APIs of asyncio are elegant and easier to understand in
my opinion.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>